pyproject.toml
[project]
name = "document-loader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.7.4",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    "numpy>=1.26.0,<2.0",
    "pandas>=2.0.0",
    "pymupdf>=1.23.8",
    "pymupdf4llm>=0.0.9",
    "python-docx>=0.8.11",
    "pyyaml>=6.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "typing-extensions>=4.0.0",
]
